refactor: organize components into proper feature folders
- Move SelectedDateSchedule to calendar/ (only used in calendar)
- Move RecentTranslation to home/ (only used in home page)
- Move UserPreference to profile/ (only used in profile pages)
- Create shared/ folder for cross-feature components:
  - UpcomingEvents (used in both home and calendar)
  - Button and Header (used across multiple features)
- Update all import paths to reflect new structure
- Better organization: feature-specific vs shared components
